requirements.txt
requests==2.31.0
watchdog==3.0.0
pywin32==306
win10toast==0.9.0
python-dotenv==1.0.1 